feature: add landing page layout
